Emo_backend/requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
joblib
pandas
numpy
nltk
scipy
python-multipart
python-dotenv
nltk
